+++ src/index.js
@@ -1,3 +1,8 @@
+import ConnectionGroup from './connection-group';
 import database from './database/root';
+import Stream from './stream';
+
+export { Graph, Node } from 'graph-crdt';
+export { Stream, ConnectionGroup };
 
 export default database;